# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CursoDisciplinaCreateDTO:
    properties:
      curso_unique_id:
        type: integer
      disciplina_unique_id:
        type: integer
    required:
    - curso_unique_id
    - disciplina_unique_id
    type: object
  dtos.CursoDisciplinaUpdateDTO:
    properties:
      curso_unique_id:
        type: integer
      disciplina_unique_id:
        type: integer
    type: object
...
  dtos.DisciplinaCreateDTO:
    properties:
      nome:
        type: string
      professor_unique_id:
        type: integer
    required:
    - nome
    - professor_unique_id
    type: object
...
  dtos.DisciplinaUpdateDTO:
    properties:
      nome:
        type: string
      professor_unique_id:
        type: integer
    type: object
...
  models.Aluno:
    properties:
      cpf:
        type: string
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cpf
    - nome
    type: object
  models.Curso:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      faculdade:
        $ref: '#/definitions/models.Faculdade'
      faculdade_unique_id:
        type: integer
      nome:
        type: string
      unique_id:
        type: integer
    type: object
  models.CursoDisciplina:
    properties:
      curso:
        $ref: '#/definitions/models.Curso'
      curso_unique_id:
        type: integer
      disciplina:
        $ref: '#/definitions/models.Disciplina'
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      disciplina_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.Disciplina:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      nome:
        type: string
      professor:
        $ref: '#/definitions/models.Professor'
      professor_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.DisciplinaMatricula:
    properties:
      aluno:
        $ref: '#/definitions/models.Aluno'
      aluno_unique_id:
        type: integer
      curso_disciplina:
        $ref: '#/definitions/models.CursoDisciplina'
      curso_disciplina_unique_id:
        type: integer
      data_matricula:
        type: string
      unique_id:
        type: integer
    type: object
  models.Faculdade:
    properties:
      cnpj:
        type: string
      cursos:
        items:
          $ref: '#/definitions/models.Curso'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cnpj
    - nome
    type: object
  models.Professor:
    properties:
      disciplinas:
        items:
          $ref: '#/definitions/models.Disciplina'
        type: array
      formacao:
        type: string
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - formacao
    - nome
    type: object
host: localhost:5000
info:
  contact:
    name: Tathiane Souza
  description: Esta é a documentação API em Golang
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Faculdade passei direto
  version: "1.0"
paths:
...
  /curso-disciplina:
    get:
      consumes:
      - application/json
      description: Busca todos os cursoDisciplinas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CursoDisciplina'
      summary: Buscar todos os cursoDisciplionas
      tags:
      - Controller CursoDisciplina
    post:
      consumes:
      - application/json
      description: Cria um cursoDisciplionas
      parameters:
      - description: cursoDiscipliona
        in: body
        name: cursoDiscipliona
        required: true
        schema:
          $ref: '#/definitions/dtos.CursoDisciplinaCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CursoDisciplina'
      summary: Cria um cursoDisciplionas
      tags:
      - Controller CursoDisciplina
  /curso-disciplina/{id}:
    delete:
      consumes:
      - application/json
      description: 'Deleta um cursoDiscipliona, os campos que devem ser enviados são:
        curso_unique_id e "disciplina_unique_id'
      parameters:
      - description: UniqueID do cursoDiscipliona
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um cursoDiscipliona
      tags:
      - Controller CursoDisciplina
    get:
      consumes:
      - application/json
      description: Busca cursoDisciplina por id
      parameters:
      - description: UniqueID do cursoDisciplina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CursoDisciplina'
      summary: Busca cursoDisciplina por id
      tags:
      - Controller CursoDisciplina
    put:
      consumes:
      - application/json
      description: 'Edita um cursoDiscipliona, os campos que podem ser enviados são:
        curso_unique_id e "disciplina_unique_id'
      parameters:
      - description: UniqueID do cursoDiscipliona
        in: path
        name: id
        required: true
        type: integer
      - description: cursoDiscipliona
        in: body
        name: cursoDiscipliona
        required: true
        schema:
          $ref: '#/definitions/dtos.CursoDisciplinaUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CursoDisciplina'
      summary: Edita um cursoDiscipliona
      tags:
      - Controller CursoDisciplina
...
  /disciplina:
    get:
      consumes:
      - application/json
      description: Busca todas as disciplinas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Disciplina'
      summary: Buscar todas as disciplinas
      tags:
      - Controller Disciplina
    post:
      consumes:
      - application/json
      description: 'Cria uma disciplina, os campos que devem ser enviados são: nome
        e professor_unique_id'
      parameters:
      - description: disciplina
        in: body
        name: disciplina
        required: true
        schema:
          $ref: '#/definitions/dtos.DisciplinaCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Disciplina'
      summary: Cria uma disciplina
      tags:
      - Controller Disciplina
...
  /disciplina/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta uma disciplina
      parameters:
      - description: UniqueID da disciplina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta uma disciplina
      tags:
      - Controller Disciplina
    get:
      consumes:
      - application/json
      description: Busca disciplina por id
      parameters:
      - description: UniqueID do disciplina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Disciplina'
      summary: Busca disciplina por id
      tags:
      - Controller Disciplina
    put:
      consumes:
      - application/json
      description: 'Edita uma disciplina, os campos que podem ser enviados são: nome
        e professor_unique_id'
      parameters:
      - description: UniqueID da disciplina
        in: path
        name: id
        required: true
        type: integer
      - description: disciplina
        in: body
        name: disciplina
        required: true
        schema:
          $ref: '#/definitions/dtos.DisciplinaUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Disciplina'
      summary: Edita uma disciplina
      tags:
      - Controller Disciplina
  /disciplina/aluno/{id}:
    get:
      consumes:
      - application/json
      description: Realiza um filtro para obter quais disciplinas um determinado aluno
        cursa
      parameters:
      - description: UniqueID do aluno
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Disciplina'
      summary: Filtra disciplinas que um determinado aluno cursa
      tags:
      - Controller Disciplina
...
swagger: "2.0"